Sipalaya/Sipalaya_backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
moviepy==1.0.3
python-decouple==3.8
Requests==2.31.0
